Add support for batch translation and Google Translate API v3
In this update, support for batch translation and Google Translate API v3 was added in the django_restful_translator application. The changes involve modifications to the translate_text function, the addition of the get_batches function in utils.py, and the update of arguments for the drt_translate_models command. The package version was also updated from 0.7.0 to 0.8.0.